nukomeet site

+++ src/backend/helpers.cpp
@@ -97,6 +97,6 @@ void processMemUsage(double& vm_usage, double& resident_set)
     }
 
     long page_size_kb = sysconf(_SC_PAGE_SIZE) / 1024; // in case x86-64 is configured to use 2MB pages
-    vm_usage = vsize / 1024.0;
+    vm_usage = (double) vsize / 1024.0;
     resident_set = rss * page_size_kb;
 }
